git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@343 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ library/ctlibreader/ct_reader/abstract/ct_abstractreader.h
@@ -70,8 +70,8 @@ public:
     // For buffer including multiples files
     struct FileBuffer {
         QString       filename;
-        size_t        nPoints;
-        QList<size_t> indexList;
+        qint64        nPoints;
+        QList<qint64> indexList;
     };
 
     /**
